move to json and tracing for easier debugging

+++ src/index/lock.rs
@@ -1,15 +1,16 @@
+use async_log::span;
 use std::io::Write;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 use std::time::Duration;
 use std::{fs, thread};
 
-use atomicwrites::AtomicFile;
-use atomicwrites::OverwriteBehavior::DisallowOverwrite;
 use uuid::Uuid;
 
 use glob::{glob, Paths};
 
+use anyhow::Context;
 use anyhow::Result;
+use std::fs::File;
 
 pub fn release_lock(path: &Path, lock_id: Uuid) -> Result<()> {
     let lock_file_path = lock_file_path(path, lock_id);
@@ -17,23 +18,26 @@ pub fn release_lock(path: &Path, lock_id: Uuid) -> Result<()> {
     Ok(())
 }
 
-fn delete_lock_file(lock_file_path: String) -> Result<()> {
-    fs::remove_file(lock_file_path.clone())?;
+fn delete_lock_file(lock_file_path: PathBuf) -> Result<()> {
+    if lock_file_path.exists() {
+        fs::remove_file(lock_file_path.clone())?;
+    }
     Ok(())
 }
 
 pub fn acquire_lock(lock_id: Uuid, index_directory: &Path) -> Result<()> {
-    create_lock_file(lock_id, index_directory)?;
     wait_to_have_sole_lock(lock_id, index_directory)?;
     Ok(())
 }
 
 pub fn wait_to_have_sole_lock(lock_id: Uuid, index_directory: &Path) -> Result<()> {
-    while !sole_lock(lock_id, index_directory)? {
-        release_lock(index_directory, lock_id)?;
-        thread::sleep(Duration::from_millis(u64::from(rand::random::<u8>())));
-        create_lock_file(lock_id, index_directory)?;
-    }
+    span!("waiting for sole lock to be {}", lock_id, {
+        while !sole_lock(lock_id, index_directory)? {
+            release_lock(index_directory, lock_id)?;
+            thread::sleep(Duration::from_millis(u64::from(rand::random::<u8>())));
+            create_lock_file(lock_id, index_directory)?;
+        }
+    });
     Ok(())
 }
 
@@ -41,7 +45,7 @@ pub fn sole_lock(lock_id: Uuid, index_directory: &Path) -> Result<bool> {
     let my_lock_file_path = lock_file_path(index_directory, lock_id);
     let mut locks = all_locks(index_directory)?;
     let only_my_locks = locks.all(|path| match path {
-        Ok(path) => path.to_string_lossy() == my_lock_file_path,
+        Ok(path) => path.to_string_lossy() == my_lock_file_path.to_string_lossy(),
         Err(_) => false,
     });
     Ok(only_my_locks)
@@ -53,13 +57,14 @@ fn all_locks(index_directory: &Path) -> Result<Paths> {
 
 fn create_lock_file(lock_id: Uuid, index_directory: &Path) -> Result<()> {
     let lock_file_path = lock_file_path(index_directory, lock_id);
-    let file = AtomicFile::new(&lock_file_path, DisallowOverwrite);
-    file.write(|f| f.write_all(lock_id.as_bytes()))?;
+    let mut file = File::create(&lock_file_path).context("creating lock file")?;
+    file.write_all(lock_id.as_bytes()).context("create lock file")?;
     Ok(())
 }
 
-fn lock_file_path(path: &Path, lock_id: Uuid) -> String {
-    format!("{}/{}.lock", path.to_string_lossy(), lock_id)
+fn lock_file_path(path: &Path, lock_id: Uuid) -> PathBuf {
+    let path_text = &format!("{}/{}.lock", path.to_string_lossy(), lock_id);
+    Path::new(path_text).to_path_buf()
 }
 
 fn locks_glob(path: &Path) -> String {
